Identify SRAM chip when flashing Genesis CFI.
In a board with 1 flash chip and SRAM, the SRAM is always enabled so attempting
to detect it as CFI will fail.

This commit adds some simple code that attempts to modify address 0x0 of the
chip to detect if the chip is ROM or SRAM. If the value changes upon writing to
it, it will assume it's SRAM and ignore it when flashing.